addressing pr comments about naming, component usage and button disabling
